+++ 96/EduBox/lib/Authenticate/LoginScreen.dart
@@ -5,9 +5,10 @@ import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 
 class LoginScreen extends StatelessWidget {
-  Future<String> signInWithGoogle() async {
-    final FirebaseAuth _auth = FirebaseAuth.instance;
-    final GoogleSignIn googleSignIn = GoogleSignIn();
+  final GoogleSignIn googleSignIn = GoogleSignIn();
+  final FirebaseAuth _auth = FirebaseAuth.instance;
+
+  Future signInWithGoogle() async {
     final GoogleSignInAccount googleSignInAccount = await googleSignIn.signIn();
     final GoogleSignInAuthentication googleSignInAuthentication =
         await googleSignInAccount.authentication;
@@ -15,114 +16,117 @@ class LoginScreen extends StatelessWidget {
       accessToken: googleSignInAuthentication.accessToken,
       idToken: googleSignInAuthentication.idToken,
     );
+    _auth.signInWithCredential(credential).then((authResult) {
+      final User user = authResult.user;
+      if (user != null) {
+        final User currentUser = _auth.currentUser;
+        assert(user.uid == currentUser.uid);
 
-    final UserCredential authResult =
-        await _auth.signInWithCredential(credential);
-    final User user = authResult.user;
-    if (user != null) {
-      assert(!user.isAnonymous);
-      assert(await user.getIdToken() != null);
-
-      final User currentUser = _auth.currentUser;
-      assert(user.uid == currentUser.uid);
-
-      bool userExists = (await FirebaseFirestore.instance
-              .collection('User')
-              .doc(user.uid)
-              .get())
-          .exists;
-      if (!userExists) {
-        FirebaseFirestore.instance.collection('User').doc(user.uid).set({
-          'Address': '',
-          'Avatar': user.photoURL,
-          'Birth': Timestamp.fromDate(DateTime(1960)),
-          'Email': user.email,
-          'Gender': '',
-          'Name': user.displayName,
-          'PhoneNumber': '',
-        });
-      } else {
-        FirebaseFirestore.instance.collection('User').doc(user.uid).update({
-          'Avatar': user.photoURL,
-          'Email': user.email,
-          'Name': user.displayName,
+        FirebaseFirestore.instance
+            .collection('User')
+            .doc(user.uid)
+            .get()
+            .then((value) {
+          if (!value.exists) {
+            FirebaseFirestore.instance.collection('User').doc(user.uid).set({
+              'Address': '',
+              'Avatar': user.photoURL,
+              'Birth': Timestamp.fromDate(DateTime(1960)),
+              'Email': user.email,
+              'Gender': '',
+              'Name': user.displayName,
+              'PhoneNumber': '',
+            });
+          } else {
+            FirebaseFirestore.instance.collection('User').doc(user.uid).update({
+              'Avatar': user.photoURL,
+              'Email': user.email,
+              'Name': user.displayName,
+            });
+          }
         });
       }
-
-      return '';
-    }
-    return null;
+    });
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: [
-          Positioned(
-            top: 95,
-            right: MediaQuery.of(context).size.width/2-100,
-            child: Container(
-              height: 200,
-              width: 200,
-              child: MyAnimatedIcon(),
-            ),
-          ),
-          Center(
-            child: GestureDetector(
-              onTap: () {
-                signInWithGoogle();
-              },
-              child: Container(
-                height: 100,
-                width: MediaQuery.of(context).size.width,
-                margin: EdgeInsets.symmetric(horizontal: 30),
-                decoration: BoxDecoration(
-                  color: Colors.blue,
-                  borderRadius: BorderRadius.circular(15),
-                  boxShadow: [
-                    BoxShadow(
-                      blurRadius: 15,
-                      offset: Offset(4.5, 4.5),
-                      color: Colors.black26,
-                    ),
-                  ],
-                ),
-                child: Row(
+    return SafeArea(
+      child: Scaffold(
+        body: StreamBuilder(
+          stream: googleSignIn.onCurrentUserChanged,
+          builder: (context, snapshot) {
+            if (snapshot.hasError)
+              return Scaffold(body: Center(child: CircularProgressIndicator()));
+            if (snapshot.hasData)
+              return Scaffold(
+                body: Center(child: LinearProgressIndicator()),
+              );
+            else
+              return Center(
+                child: Column(
                   mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
                   children: [
                     Container(
-                      padding: EdgeInsets.all(4),
-                      decoration: BoxDecoration(
-                        color: Colors.white,
-                        boxShadow: [
-                          BoxShadow(
-                            blurRadius: 3,
-                            offset: Offset(4, 4),
-                            color: Colors.black26,
-                          ),
-                        ],
-                      ),
-                      child: Image.asset('lib/assets/google.png'),
-                      height: 45,
-                      width: 45,
+                      height: MediaQuery.of(context).size.width / 2,
+                      width: MediaQuery.of(context).size.width / 2,
+                      child: MyAnimatedIcon(),
                     ),
-                    Container(
-                      margin: EdgeInsets.symmetric(horizontal: 20),
-                      child: Text(
-                        'Đăng nhập với google',
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontSize: 20,
+                    GestureDetector(
+                      onTap: signInWithGoogle,
+                      child: Container(
+                        height: 100,
+                        width: MediaQuery.of(context).size.width / 5 * 4,
+                        decoration: BoxDecoration(
+                          color: Colors.blue,
+                          borderRadius: BorderRadius.circular(15),
+                          boxShadow: [
+                            BoxShadow(
+                              blurRadius: 15,
+                              offset: Offset(4.5, 4.5),
+                              color: Colors.black26,
+                            ),
+                          ],
+                        ),
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Container(
+                              padding: EdgeInsets.all(4),
+                              decoration: BoxDecoration(
+                                color: Colors.white,
+                                boxShadow: [
+                                  BoxShadow(
+                                    blurRadius: 3,
+                                    offset: Offset(4, 4),
+                                    color: Colors.black26,
+                                  ),
+                                ],
+                              ),
+                              child: Image.asset('lib/assets/google.png'),
+                              height: 45,
+                              width: 45,
+                            ),
+                            Container(
+                              margin: EdgeInsets.symmetric(horizontal: 20),
+                              child: Text(
+                                'Đăng nhập với google',
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 20,
+                                ),
+                              ),
+                            ),
+                          ],
                         ),
                       ),
                     ),
                   ],
                 ),
-              ),
-            ),
-          ),
-        ],
+              );
+          },
+        ),
       ),
     );
   }
