+++ 96/EduBox/lib/Home/Button.dart
@@ -29,7 +29,6 @@ class Button extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // TODO: implement build
     return Container(
       height: 130,
       width: 110,
