+++ 96/EduBox/lib/LandingPage.dart
@@ -7,22 +7,10 @@ class LandingPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<User>(
-      stream: FirebaseAuth.instance.authStateChanges(),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.active) {
-          User user = snapshot.data;
-          //print(user);
-          return user != null? HomeInterface():
-          LoginScreen();
-        }
-        else {
-          return Scaffold(
-            body: Center(
-              child: CircularProgressIndicator(),
-            ),
-          );
-        }
-      },
-    );
+        stream: FirebaseAuth.instance.authStateChanges(),
+        builder: (context, snapshot) {
+          if (snapshot.hasData) return HomeInterface();
+          return LoginScreen();
+        });
   }
 }
