+++ 96/EduBox/pubspec.yaml
@@ -28,7 +28,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons:
   intl: ^0.16.1
   clay_containers:
   firebase_core:
